Boards and Sprites
Added red and blue tilemap prefabs.  Drag and drop into scene for them to work, should load in fully with a board script and collider for the grid.

Also added new ships into the sprites folder.  Should be able to drag and drop them directly to the scene where they should also already be the correct size, but can only do this in scene editor as of now and not actual gameplay.

Squares on the grid are 32x32 pixels big. The ships are the Destroyer (two hp, 32x64 pixels), Sub and Cruiser (three hp, 32x96 pixels), a battleship (four hp, 32x128 pixels), and a carrier (five hp, 32x160)

Finally, attempted at making a drag and drop from a tutorial but couldn't figure out how to actually make it run.  I left the code that won't crash the program but won't actually do anything under the dragDrop script.